Removed unnecessary calls to list blobs in container and moved file to sdk/storage/azure-storage-blobs/samples
